+++ src/GLWpfControl/GLWpfControl.csproj
@@ -40,18 +40,10 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="GLWpfControl.xaml.cs">
-      <DependentUpon>GLWpfControl.xaml</DependentUpon>
-    </Compile>
+    <Compile Include="GLWpfControl.cs" />
     <Compile Include="GLWpfControlRenderer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="GLWpfControlSettings.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <Page Include="GLWpfControl.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
   </ItemGroup>
   <ItemGroup>
     <Content Include="OpenTK.dll.config">
@@ -59,6 +51,7 @@
     </Content>
     <Content Include="paket.references" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Choose>
     <When Condition="$(TargetFrameworkIdentifier) == '.NETFramework' And $(TargetFrameworkVersion) == 'v4.5.2'">
